+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - `robots` plugin: support Content-Signal rule
 - `feed` plugin: Allow `html` extension for pretty URL [#839]
 - `feed` plugin: New `info.self` option to customize the `self` url [#839]
+- `image_size` plugin: Support for remote files.
 
 ### Changed
 - `robots` and `sitemap` plugins: the generated `robots.txt` content will always end with a newline